Backend/CVModules/CVExtractionCloudTrigger/requirements.txt
functions-framework==3.*
requests>=2.32.5
google-cloud-storage>=2.10.0
google-cloud-firestore==2.14.0
